backend/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[project]
name = "moni-backend"
version = "0.1.0"
description = "MONI monitoring API"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "pydantic-settings>=2.0",
  "python-multipart>=0.0.9",
  "email-validator>=2.0",
  "celery[redis]>=5.4",
  "redis>=5.0",
  "httpx>=0.27",
  "python-whois>=0.9.4",
  "python-jose[cryptography]>=3.3",
  "bcrypt>=4.1",
  "psycopg[binary]>=3.2",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.8",
  "pytest>=8.0",
  "pytest-asyncio>=0.24",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
